+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Use `child_class` property in the populate method
 * Add filter logic to widgets for getting specific values
 * Add abstract `Partial` widget class, to easily create a new widget with a pre-defined layout
+* Wait for widgets to render their contents before starting editor
 
 ## 0.1.5 (2022-07-14)